Refactor folder structure and update database connection
