+++ .gitignore
@@ -2,5 +2,4 @@ node_modules/
 .env
 dist/
 output.log
-
 coverage/
