+++ src/__test__/globalConfig.ts
@@ -1,5 +1,5 @@
-import request from "supertest";
-import app from "../index";
+import request from 'supertest';
+import app from '../app';
 
 export async function beforeAllHook() {
   jest.setTimeout(50000);
